simple procedural skybox
